requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.7
django-bootstrap-v5==1.0.1
django-tinymce==3.3.0
gunicorn==20.0.4
importlib-metadata==1.7.0
pystache==0.5.4
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
static3==0.7.0
zipp==3.4.1
psycopg2>=2.6.1